pyproject.toml
[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.vulture]
min_confidence = 70
sort_by_size = true

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
indent='    '
sections = ['FUTURE','STDLIB','THIRDPARTY','FIRSTPARTY','LOCALPROJECT','LOCALFOLDER']
default_section = 'THIRDPARTY'
known_standard_library = []
known_localproject = ['app','tests']

[tool.pytest.ini_options]
env_override_existing_values = 0
env_files = [".env", ".env.test"]
filterwarnings = [
    'ignore'
]


[tool.poetry]
name = "billing_system_test_task"
version = "1.0.1"
description = "Test task with implementation of simple billing system"
authors = ["Vadim Sokoltsov <vforvad@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = "^0.15.0"
python-dotenv = "^0.19.1"
alembic = "^1.7.4"
databases = "^0.5.3"
asyncpg = "^0.24.0"
pydantic = "^1.8.2"
psycopg2-binary = "^2.9.1"
email-validator = "^1.1.3"
uvloop = "^0.16.0"
pylint = "^2.11.1"
autopep8 = "^1.5.7"

[tool.poetry.dev-dependencies]
ipython = "^7.28.0"
pycodestyle = "^2.8.0"
flake8 = "^4.0.1"
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
factory-boy = "^3.2.0"
pytest-factoryboy = "^2.1.0"
pytest-dotenv = "^0.5.2"
pyint = "^1.1.0"
mypy = "^0.910"
black = "^21.9b0"
sqlalchemy-stubs = "^0.4"
pytest-cov = "^3.0.0"
isort = "^5.9.3"
httpx = "^0.20.0"
requests = "^2.26.0"
bandit = "^1.7.0"
mock = "^4.0.3"
Faker = "^9.7.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
